graphql changes

+++ src/main/java/org/problems/graphql/simple/model/Book.java
@@ -7,5 +7,5 @@ import java.math.BigDecimal;
 @Data
 public class Book {
     private String name;
-    private BigDecimal cost;
+    private Integer cost;
 }
